--- src/components/Account/Account.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-//components
-import HeaderNav from '../Header/HeaderNav/HeaderNav';
-
-//styles
-import './AccountStyles';
-
-const Account = () => {
-  return (
-    <>
-      <HeaderNav />
-    </>
-  )
-}//end Account
-
-
-export default Account;
