--- src/components/VolunteerProfile/VolunteerProfile.js
@@ -1,18 +1,0 @@
-// - `username`: String
-// - `volunteerName`: String
-// - `phoneNumber` 
-
-import React from 'react';
-
-//styles
-import './VolunteerProfileStyles';
-
-const VolunteerProfile = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}// end VolunteerProfile
-
-export default VolunteerProfile;
