--- src/components/VolunteerProfile/VolunteerProfileStyles.js
@@ -1,1 +1,0 @@
-import styled from 'styled-components';
